Move remaining helpers into app.rb
